moved constraint helpers
