+++ src/generators/go/constrainer/ModelNameConstrainer.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
-import { NO_NUMBER_START_CHAR, NO_EMPTY_VALUE} from '../../../helpers/Constraints';
+import { NO_NUMBER_START_CHAR, NO_EMPTY_VALUE} from '../../../helpers/CommonConstraints';
 import { FormatHelpers } from '../../../helpers';
 
 export type ModelNameConstraints = {
