+++ src/generators/go/constrainer/PropertyKeyConstrainer.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import { ConstrainedObjectModel, ObjectModel } from '../../../models';
-import { NO_NUMBER_START_CHAR, NO_DUPLICATE_PROPERTIES, NO_EMPTY_VALUE} from '../../../helpers/Constraints';
+import { NO_NUMBER_START_CHAR, NO_DUPLICATE_PROPERTIES, NO_EMPTY_VALUE} from '../../../helpers/CommonConstraints';
 import { FormatHelpers } from '../../../helpers';
 
 export type PropertyKeyConstraintOptions = {
